refactor(core): restructure server and service classes, improve modularity WIP
