+++ mlox/session.py
@@ -2,9 +2,6 @@ from dataclasses import dataclass, field
 
 from mlox.infra import Infrastructure
 from mlox.secret_manager import TinySecretManager
-from mlox.utils import dataclass_to_dict, dict_to_dataclass
-from mlox.server import AbstractServer
-from mlox.service import AbstractService
 
 
 @dataclass
@@ -22,14 +19,14 @@ class MloxSession:
         self.load_infrastructure()
 
     def save_infrastructure(self) -> None:
-        infra_dict = dataclass_to_dict(self.infra)
+        infra_dict = self.infra.to_dict()
         self.secrets.save_secret("MLOX_CONFIG_INFRASTRUCTURE", infra_dict)
 
     def load_infrastructure(self) -> None:
         infra_dict = self.secrets.load_secret("MLOX_CONFIG_INFRASTRUCTURE")
-        if infra_dict and isinstance(infra_dict, dict):
-            self.infra = dict_to_dataclass(
-                infra_dict, hooks=[AbstractServer, AbstractService]
-            )
-        else:
-            self.infra = Infrastructure()
+        print(infra_dict)
+        if not infra_dict:
+            raise ValueError("No infrastructure data found in secrets.")
+        if not isinstance(infra_dict, dict):
+            raise ValueError("Infrastructure data is not in the expected format.")
+        self.infra = Infrastructure.from_dict(infra_dict)
