+++ mlox/view/login.py
@@ -1,10 +1,11 @@
 import os
 import streamlit as st
 
+from mlox.services.tsm.service import TSMService
 from mlox.session import MloxSession
+from mlox.config import load_all_server_configs, load_config
 from mlox.infra import Infrastructure
 from mlox.utils import dataclass_to_dict, save_to_json
-from mlox.view.utils import form_add_server
 
 
 def login():
@@ -36,25 +37,27 @@ def new_project():
             value=os.environ.get("MLOX_CONFIG_PASSWORD", ""),
             type="password",
         )
-        ip, port, root, pw, config = form_add_server()
-        submitted = st.form_submit_button("New")
-        if submitted:
-            bundle = Infrastructure().add_server(
-                config,
-                {
-                    "${MLOX_IP}": ip,
-                    "${MLOX_PORT}": str(port),
-                    "${MLOX_ROOT}": root,
-                    "${MLOX_ROOT_PW}": pw,
-                },
-            )
+        configs = load_all_server_configs("./stacks")
+        config = st.selectbox(
+            "System Configuration",
+            configs,
+            format_func=lambda x: f"{x.name} {x.version} - {x.description_short}",
+            help="Please select the configuration that matches your server.",
+        )
+        params = dict()
+        infra = Infrastructure()
+        setup_func = config.instantiate_ui("setup")
+        if setup_func:
+            params = setup_func(infra)
+        if st.form_submit_button("Submit", type="primary"):
+            bundle = infra.add_server(config, params)
             if not bundle:
                 st.error(
                     "Something went wrong. Check server credentials and try again."
                 )
             else:
                 with st.spinner("Initializing server, writing keyfile, etc..."):
-                    bundle.initialize()  # initialize server (add mlox user, ssh keys, update etc)
+                    bundle.server.setup()
                     server_dict = dataclass_to_dict(bundle.server)
                     save_to_json(server_dict, f"./{username}.key", password, True)
 
@@ -62,7 +65,11 @@ def new_project():
             try:
                 ms = MloxSession(username, password)
                 if ms.secrets.is_working():
-                    ms.infra.bundles.append(bundle)
+                    ms.infra = infra
+                    config = load_config("./stacks", "/tsm", "mlox.tsm.yaml")
+                    ms.infra.add_service(bundle.server.ip, config, {})
+                    ms.infra.get_service(TSMService.__class__.__name__).pw = password
+                    print(ms.infra.get_service(TSMService.__class__.__name__).pw)
                     bundle.tags.append("mlox-secrets")
                     ms.save_infrastructure()
                     st.session_state["mlox"] = ms
