--- stacks/ubuntu/mlox-server.ubuntu-test.yaml
@@ -1,45 +1,0 @@
-name: Ubuntu-Test
-versions: [24.04]
-maintainer: Your Name
-description: |
-  Ubuntu is a popular Linux distribution based on Debian.
-  It is known for its ease of use, stability, and strong community support.
-  This YAML file defines the configuration for the base Ubuntu stack which is necessary for the
-  installation and management of MLOX stacks.
-links:
-  project: https://ubuntu.com/
-  news: https://ubuntu.com/
-  security: https://ubuntu.com/
-  documentation: https://ubuntu.com/
-  changelog: https://ubuntu.com/
-build:
-  class_name: mlox.server.ConfigurableServer
-  params:
-    script_test: [
-      "ls -l",
-      ["sys_disk_free"],
-      # ["sys_add_user", "test_user", "test_pw", True, True],
-      # [False, "mkdir -p ~/.ssh; chmod 700 ~/.ssh", "Add SSH directory"],
-      [False, "echo 'Hello, world!' > /home/test_user/hello.txt", "Add hello.txt"],
-      "echo 'Hello, world!'"]
-    script_generate_ssh_keys: [
-      [False, "mkdir -p ~/.ssh; chmod 700 ~/.ssh", "Add SSH directory"],
-      [False, "cd /home/${MLOX_USER}/.ssh; rm id_rsa*; ssh-keygen -b 4096 -t rsa -f id_rsa -N ${MLOX_SSH_PW}", "Generate SSH key"],
-      ["fs_append_line", "/home/${MLOX_USER}/.ssh/authorized_keys", "${MLOX_SSH_PUB_KEY}", "Add SSH public key"],
-    ]
-    script_setup: [
-      ["sys_add_user", "${MLOX_USER}", "${MLOX_PW}", True, True],  # add user with sudo privileges
-      [False, "mkdir -p ~/.ssh; chmod 700 ~/.ssh", "Add SSH directory"],
-      [False, "cd /home/${MLOX_USER}/.ssh; rm id_rsa*; ssh-keygen -b 4096 -t rsa -f id_rsa -N ${MLOX_SSH_PW}", "Generate SSH key"],
-      ["fs_append_line", "/home/${MLOX_USER}/.ssh/authorized_keys", "${MLOX_SSH_PUB_KEY}", "Add SSH public key"],
-      ["fs_find_and_replace", "/etc/ssh/sshd_config", "#PasswordAuthentication yes", "PasswordAuthentication no"],
-      ["fs_find_and_replace", "/etc/ssh/sshd_config", "X11Forwarding yes", "X11Forwarding no"],
-      ["fs_find_and_replace", "/etc/ssh/sshd_config", "AllowTcpForwarding yes", "AllowTcpForwarding no"],
-      [False, "systemctl restart ssh", "Restart SSH"],
-    ]
-    script_update: [
-      [True, "dpkg --configure -a", "Update packages"],
-      [True, "apt-get update", "Update packages"],
-      [True, "apt-get -y upgrade", "Upgrade packages"],
-    ]
-
