+++ stacks/ubuntu/mlox-server.ubuntu.docker.yaml
@@ -1,6 +1,8 @@
-name: Ubuntu
-versions: [24.04, 22.04, 20.04]
+name: Ubuntu (Docker)
+version: 24.04 
+variants: [24.04, 22.04, 20.04]
 maintainer: Your Name
+description_short: Ubuntu with Docker backend.
 description: |
   Ubuntu is a popular Linux distribution based on Debian.
   It is known for its ease of use, stability, and strong community support.
@@ -12,8 +14,16 @@ links:
   security: https://ubuntu.com/
   documentation: https://ubuntu.com/
   changelog: https://ubuntu.com/
+groups:
+  git:
+  server:
+  backend:
+    docker:
+ui:
+  setup: mlox.servers.ubuntu.ui_docker.setup
+  settings: mlox.servers.ubuntu.ui_docker.settings
 build:
-  class_name: mlox.server.Ubuntu
+  class_name: mlox.servers.ubuntu.docker.UbuntuDockerServer
   params:
     ip: ${MLOX_IP}
     port: ${MLOX_PORT}
